control tests need -lm

+++ tests/du.c
@@ -208,7 +208,7 @@ int durecurse(char *path, size_t len, int *opt)
         else
 	{ 
 		lstat(path, &sb);
-		printf("%zu\t%s\n", (sb.st_blocks * 512 ) /hold.block, path); 
+		printf("%zu\t%s\n", (size_t)(sb.st_blocks * 512 ) /hold.block, path); 
 		return 0;
 	}
